NOJIRA: refactor node_module serving to be far less cumbersome using another piece of middleware.
