+++ index.js
@@ -12,7 +12,7 @@ https://raw.githubusercontent.com/fluid-project/sjrk-story-telling-server/master
 var fluid = require("infusion");
 require("kettle");
 require("./src/js/staticHandlerBase");
-// require("./src/js/nodeModuleMounter");
+require("./src/js/middleware");
 require("./src/js/dataSource");
 require("./src/js/serverSetup");
 require("./src/js/requestHandlers");
